Allow OAUTH2 authenticator to also accept bearer tokens
Add `allowBearerAuthentication` to OAuth2Authenticator so a single endpoint can
serve both a human in a browser (interactive OIDC login) and an external system
presenting an `Authorization: Bearer <jwt>` token — the machine obtains the token
via the client_credentials grant at the same IdP.

The bearer path reuses the JWT validation already in BearerOnlyAuthenticator:
its decoder, audience check, userInfo enrichment and claim->authority mapping are
extracted into a shared AbstractJwtAuthenticator that both authenticators extend.
BearerOnlyAuthenticator is unchanged in behaviour. When enabled, OAuth2Authenticator
adds oauth2ResourceServer().jwt(...) alongside oauth2Login and installs a
DelegatingAuthenticationEntryPoint so API/bearer requests get a 401 (Bearer) while
browser requests still get the login redirect. Requires issuerUri or jwkSetUri;
defaults to false so existing OAUTH2 deployments are unaffected.
